+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "babel-core": "^6.13.2",
     "babel-loader": "^6.2.4",
+    "babel-plugin-transform-object-rest-spread": "6.8.0",
     "babel-preset-es2015": "^6.13.2",
     "css-loader": "^0.23.1",
     "express": "^4.14.0",
