+++ webpack.config.js
@@ -15,7 +15,8 @@ module.exports = {
         loader: 'babel-loader',
         exclude: /node_modules/,
         query: {
-          presets: ['es2015']
+          presets: ['es2015'],
+          plugins: ["transform-es2015-destructuring", "transform-object-rest-spread"]
         }
       },
       {
